requirements.txt
django>=6.0
djangorestframework>=3.14
mssql-django>=1.4
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
